feat: implement batch size validation and automatic splitting for DA (#2506)
Now is responsability of the node to split the batch that is sent to the
DA. And the client only validates if the blob is too big (does not tries
to split it)

NOTE: There exists a case when the blob of a single block is too big, it
wont be split and will fail after 30 attempts.'

Closes: https://github.com/evstack/ev-node/issues/2498

<!--
Please read and fill out this form before submitting your PR.

Please make sure you have reviewed our contributors guide before
submitting your
first PR.

NOTE: PR titles should follow semantic commits:
https://www.conventionalcommits.org/en/v1.0.0/
-->

## Overview

<!-- 
Please provide an explanation of the PR, including the appropriate
context,
background, goal, and rationale. If there is an issue with this
information,
please provide a tl;dr and link the issue. 

Ex: Closes #<issue number>
-->

---------

Co-authored-by: Jonathan Gimeno <jgimeno@gmail.com>