fix: Modify Max DA blob size slightly (#2558)
<!--
Please read and fill out this form before submitting your PR.

Please make sure you have reviewed our contributors guide before
submitting your
first PR.

NOTE: PR titles should follow semantic commits:
https://www.conventionalcommits.org/en/v1.0.0/
-->

## Overview

The DA Max Blob size in Celestia according to
https://docs.celestia.org/how-to-guides/submit-data#maximum-blob-size is
1,973,786 bytes which is equal to 1973786/(1024*1024) = 1.882 MB. This
is a bit lower than 64*64*482 = 1974272 which is what was set in Evolve
Node, so this PR modifies this value in Evolve Node to be 64*64*481 =
1970176 instead so it's a bit lower than the DA Max Blob size instead.

Also, modified the logs to be more debug based instead of error and info
to reduce the noise in logs for users.

<!-- 
Please provide an explanation of the PR, including the appropriate
context,
background, goal, and rationale. If there is an issue with this
information,
please provide a tl;dr and link the issue. 

Ex: Closes #<issue number>
-->